requirements.txt
# General Utilities
numpy
pandas
scipy
tqdm
h5py
joblib
python-dateutil

# Machine Learning and Deep Learning Frameworks
tensorflow-macos
tensorflow-metal
keras
torchvision
tensorflow-addons

# Natural Language Processing
nltk
spacy
gensim
transformers

# Text Processing Utilities
regex
unidecode
beautifulsoup4

# Image Processing
opencv-python
Pillow
scikit-image
albumentations
imgaug

# Medical Image Processing
pydicom
SimpleITK
nibabel

# Visualization
matplotlib
seaborn
plotly
tensorboard

# Evaluation Metrics
scikit-learn

# Configuration and Environment
python-dotenv
pydantic

# Serialization and Data Formats
pickle5
hdf5storage
msgpack

# Advanced Mathematical Functions
sympy
numba

# For BERT, GPT Models
sentencepiece

# Multiprocessing and Parallelism
multiprocess